+++ internal/infrastructure/providers/accrual/accrual_test.go
@@ -1,4 +1,4 @@
-package accrualclient
+package accrual
 
 import (
 	"context"
