+++ internal/infrastructure/providers/accrual/errors.go
@@ -1,4 +1,4 @@
-package accrualclient
+package accrual
 
 import (
 	"errors"
