+++ internal/service/gophermartservice/service.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	Accrual "github.com/zaz600/go-musthave-diploma/api/accrual"
-	"github.com/zaz600/go-musthave-diploma/internal/infrastructure/providers/accrualclient"
+	"github.com/zaz600/go-musthave-diploma/internal/infrastructure/providers/accrual"
 	"github.com/zaz600/go-musthave-diploma/internal/infrastructure/repository"
 )
 
@@ -13,13 +13,13 @@ const accrualDefaultRetryInterval = 50 * time.Millisecond
 type GophermartService struct {
 	repo repository.RepoRegistry
 
-	accrualClient        *accrualclient.Client
+	accrualClient        *accrual.Client
 	accrualRetryInterval time.Duration
 }
 
 func New(accrualAPIClient Accrual.ClientWithResponsesInterface, opts ...Option) *GophermartService {
 	s := &GophermartService{
-		accrualClient:        accrualclient.New(accrualAPIClient),
+		accrualClient:        accrual.New(accrualAPIClient),
 		accrualRetryInterval: accrualDefaultRetryInterval,
 	}
 	for _, opt := range opts {
